Add mailman3 backend \o/

+++ src/update_mailman.sh
@@ -41,40 +41,35 @@ echo $$ > "$LOCK_FILE"
 
 # List the lists to CREATE
 mysql_query "SELECT id,list, name, domain, owner, password FROM mailman WHERE mailman_action='CREATE';"|while read id list name domain owner password ; do
-  if [ -d "/var/lib/mailman/lists/$name" ]
+  if [ -d "/var/lib/mailman3/lists/$name.$domain" ]
     then
     mysql_query "UPDATE mailman SET password='', mailman_result='This list already exist', mailman_action='OK' WHERE id='$id';"
     else
       # Create the list : 
-      sudo -u list /usr/lib/mailman/bin/newlist -q "$list@$domain" "$owner" "$password"
+      sudo -u list /usr/lib/mailman3/bin/mailman create -q $list@$domain -o $owner
       if [ "$?" -eq "0" ]
       then
         mysql_query "UPDATE mailman SET password='', mailman_result='', mailman_action='OK' WHERE id='$id';"
-        sudo -u list /usr/lib/mailman/bin/withlist -q -l -r set_url_alternc "$name" "$MAILMAN_URL"
-      if [ "$?" -ne "0" ]
-        # SetURL the list with the default fqdn to start: 
-      then
-        mysql_query "UPDATE mailman SET mailman_result='A fatal error happened when changing the list url', mailman_action='OK' WHERE id='$id';"
-      else
-        mysql_query "UPDATE mailman SET mailman_action='OK' WHERE id='$id';"
       fi
     fi
-  fi
 done
 
 # List the lists to DELETE
-mysql_query "SELECT id, list, name, domain FROM mailman WHERE mailman_action='DELETE';"|while read id list name domain ; do
-  if [ ! -d "/var/lib/mailman/lists/$name" ]
+mysql_query "SELECT id, list, name, domain, mailman_version FROM mailman WHERE mailman_action='DELETE';"|while read id list name domain mailman_version; do
+  if [ ! -d "/var/lib/mailman3/lists/$name.$domain" -a ! -d "/var/lib/mailman/lists/$name" ]
     then
     mysql_query "UPDATE mailman SET mailman_result='This list does not exist', mailman_action='OK' WHERE id='$id';"
     else
-# Delete the list : 
     mysql_query "UPDATE mailman SET mailman_action='DELETING' WHERE id='$id';"
-    sudo -u list /usr/lib/mailman/bin/rmlist "$name"
+    if (( $(echo "$mailman_version >= 3" |bc -l) ))
+    then
+        sudo -u list /usr/lib/mailman3/bin/mailman remove "$name@$domain"
+    else
+        sudo -u list /usr/lib/mailman/bin/rmlist "$name"
+    fi
     if [ "$?" -eq "0" ]
-      then
-      # Now delete the archives too ...
-      sudo -u list /usr/lib/mailman/bin/rmlist -a "$name"
+    then
+      # TODO: check if archives deletion is needed here
       mysql_query "DELETE FROM mailman WHERE id='$id';"
     else
       mysql_query "UPDATE mailman SET mailman_result='A fatal error happened when deleting the list', mailman_action='OK' WHERE id='$id';"
@@ -83,13 +78,12 @@ mysql_query "SELECT id, list, name, domain FROM mailman WHERE mailman_action='DE
 done
 
 
-# List the lists to PASSWORD
+# List the lists to PASSWORD (mailman 2 only)
 mysql_query "SELECT id, list, name, domain, password FROM mailman WHERE mailman_action='PASSWORD';"|while read id list name domain password ; do
   if [ ! -d "/var/lib/mailman/lists/$name" ]
     then
     mysql_query "UPDATE mailman SET mailman_result='This list does not exist', mailman_action='OK', password='' WHERE id='$id';"
     else
-# Password the list : 
     sudo -u list /usr/lib/mailman/bin/change_pw -l "$name" -p "$password"
     if [ "$?" -eq "0" ]
       then
@@ -119,53 +113,6 @@ mysql_query "SELECT id, list, name, domain FROM mailman WHERE mailman_action='GE
 done
 
 
-# List the lists to REGENERATE
-mysql_query "SELECT id, list, name, domain, url FROM mailman WHERE mailman_action='REGENERATE';"|while read id list name domain url ; do
-#non virtual lists
-if [ "$list" == "$name" ]; then
-  if [ ! -d "/var/lib/mailman/lists/$list" ]
-    then
-    mysql_query "UPDATE mailman SET mailman_result='This list does not exist', mailman_action='OK' WHERE id='$id';"
-  else
-   	sudo -u list /usr/lib/mailman/bin/withlist -q -l -r set_url_alternc "$name" "$MAILMAN_URL"
-    if [ "$?" -eq "0" ]
-      then
-      mysql_query "UPDATE mailman SET mailman_result='', mailman_action='OK' WHERE id='$id';"
-    else
-      mysql_query "UPDATE mailman SET mailman_result='A fatal error happened when changing the list url', mailman_action='OK' WHERE id='$id';"
-    fi
-  fi
-else
-#virtual lists
-  if [ ! -d "/var/lib/mailman/lists/$name" ];then
-  #virtual list just just virtualised ( /var/lib/mailman/lists/$list exists )
-    sudo -u list /usr/lib/mailman/bin/withlist -q -l -r set_url_alternc "$name" "$MAILMAN_URL"
-    if [ "$?" -eq "0" ]
-      then
-      mysql_query "UPDATE mailman SET mailman_result='', mailman_action='OK' WHERE id='$id';"
-    else
-      mysql_query "UPDATE mailman SET mailman_result='A fatal error happened when changing the list url', mailman_action='OK' WHERE id='$id';"
-    fi
-  #move the list to match its new name : 
-    mv "/var/lib/mailman/lists/$list" "/var/lib/mailman/lists/$name"
-    mv "/var/lib/mailman/archives/private/$list" "/var/lib/mailman/archives/private/$name"
-    mv "/var/lib/mailman/archives/private/$list.mbox" "/var/lib/mailman/archives/private/$name.mbox"
-    mv "/var/lib/mailman/archives/private/$name.mbox/$list.mbox" "/var/lib/mailman/archives/private/$name.mbox/$name.mbox"
-    if [ -e "/var/lib/mailman/archives/private/$name.mbox/$name.mbox" ];then
-      /usr/lib/mailman/bin/arch --wipe $name
-      if [ "$?" -eq "0" ]
-        then
-        mysql_query "UPDATE mailman SET mailman_result='', mailman_action='REGENERATE-2' WHERE id='$id';"
-      else
-        mysql_query "UPDATE mailman SET mailman_result='A fatal error happened when regenerating the list', mailman_action='OK' WHERE id='$id';"
-      fi
-    fi
-  else
-    mysql_query "UPDATE mailman SET mailman_result='', mailman_action='REGENERATE-2' WHERE id='$id';"
-  fi
-fi
-done
-
 # List the lists to SETURL
 mysql_query "SELECT id, list, name, domain, url FROM mailman WHERE mailman_action='SETURL';"|while read id list name domain url ; do
   if [ ! -d "/var/lib/mailman/lists/$name" ]
@@ -181,6 +128,20 @@ mysql_query "SELECT id, list, name, domain, url FROM mailman WHERE mailman_actio
       mysql_query "UPDATE mailman SET mailman_result='A fatal error happened when changing the list url', mailman_action='OK' WHERE id='$id';"
     fi
   fi
+done
+
+
+# List the lists to MIGRATE
+mysql_query "SELECT id, list, name, domain FROM mailman WHERE mailman_action='MIGRATE';"|while read id list name domain owner; do
+    MIGRATION_LOG_FILE="/var/log/alternc/list_migrate.log"
+    /usr/sbin/list_migrate $list@$domain >> $MIGRATION_LOG_FILE 2>> $MIGRATION_LOG_FILE
+    if [ $? -eq 0 ]
+    then
+      mysql_query "UPDATE mailman SET mailman_result='', mailman_action='OK', mailman_version=3 WHERE id='$id';"
+    else
+      echo "An error occured while migrating $list@$domain, see $MIGRATION_LOG_FILE for details"
+      mysql_query "UPDATE mailman SET mailman_result='An error happened while trying to migrate the list, please contact your admin', mailman_action='OK' WHERE id='$id';"
+    fi
 done
 
 # Delete the lock
